Refactoring to allow building for MSVC targets.

+++ Thor/drivers/f4/iwdg/hw_iwdg_driver_stm32f4.cpp
@@ -90,7 +90,7 @@ namespace Thor::Driver::IWDG
     correlates to the actual prescaler numerical divisor.
     ------------------------------------------------*/
     auto pair     = Chimera::Utilities::findInVector( prescalerRegVals, prescaler );
-    int index     = pair.second;
+    size_t index  = pair.second;
     bool validity = pair.first;
 
     if ( !validity )
